Refactored location of blog data /components/Blogs.js -> /resources/blogData.js
